requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
streamlit-drawable-canvas